ai-chatbox/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.2
python-multipart>=0.0.6
sqlalchemy>=2.0.19
google-generativeai>=0.3.0
pypdf>=3.17.0
python-dotenv>=1.0.0
